Add files via upload
This is a code written in C++ which allows for the user to add employees to an existing list of people. The user can add their name, ID number, and salary. The functions that were shown are able to: add an employee, display the employees, and remove and employee. The other functions that weren't shown in the screen shoot are to search and find an employee, and change their salary.